+++ .eslintrc.js
@@ -12,11 +12,9 @@ module.exports = {
     sourceType: 'module'
   },
   rules: {
-    'indent': ['error', 2],
     'linebreak-style': ['error', 'unix'],
     'quotes': ['error', 'single'],
     'semi': ['error', 'always'],
     'no-unused-vars': 'warn',
-    'no-console': 'warn'
   }
 };